pyproject.toml
[project]
name = "spotify-playlist-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.2",
    "spotipy>=2.26.0",
]
